# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/search": {
      "post": {
        "operationId": "AlojamientosController_search",
        "summary": "Búsqueda de alojamientos",
        "parameters": [
          {
            "name": "X-Device-Fingerprint",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchAccommodationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alojamientos encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchAccommodationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida (ProblemDetails)"
          },
          "429": {
            "description": "Demasiadas peticiones (ProblemDetails + Retry-After)"
          }
        },
        "tags": [
          "Búsqueda y Catálogo"
        ]
      }
    },
    "/api/v1/availability": {
      "post": {
        "operationId": "AlojamientosController_availability",
        "summary": "Consultar disponibilidad y precio de un alojamiento",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Disponibilidad y detalles del precio",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida (ProblemDetails)"
          }
        },
        "tags": [
          "Disponibilidad y Precios"
        ]
      }
    },
    "/api/v1/bulk-availability": {
      "post": {
        "operationId": "AlojamientosController_bulkAvailability",
        "summary": "Consultar disponibilidad múltiple de alojamientos",
        "description": "filters.meal_plan no está soportado y se ignora.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAvailabilityRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Disponibilidad para múltiples alojamientos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkAvailabilityResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Disponibilidad y Precios"
        ]
      }
    },
    "/api/v1/details": {
      "post": {
        "operationId": "AlojamientosController_getDetails",
        "summary": "Obtener detalles extendidos de los alojamientos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccommodationDetailsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detalles de los alojamientos solicitados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDetailsResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Búsqueda y Catálogo"
        ]
      }
    },
    "/api/v1/details/changes": {
      "post": {
        "operationId": "AlojamientosController_getDetailsChanges",
        "summary": "Obtener alojamientos que han cambiado desde una fecha",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetailsChangesRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lista de alojamientos modificados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailsChangesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido (ProblemDetails)"
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "alojamientos:read"
            ]
          }
        ],
        "tags": [
          "Búsqueda y Catálogo"
        ]
      }
    },
...
    "/api/v1/constants": {
      "post": {
        "operationId": "AlojamientosController_getConstants",
        "summary": "Consultar constantes del sistema (facilidades, tipos de cuartos, etc.)",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstantsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Constantes del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstantsResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Componentes Comunes"
        ]
      }
    },
    "/api/v1/reviews": {
      "post": {
        "operationId": "AlojamientosController_getReviews",
        "summary": "Obtener reseñas de alojamientos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reseñas de los alojamientos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewsResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Búsqueda y Catálogo"
        ]
      }
    },
    "/api/v1/reviews/scores": {
      "post": {
        "operationId": "AlojamientosController_getReviewsScores",
        "summary": "Obtener puntuaciones de reseñas",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewsScoresRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Puntuaciones desglosadas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewsScoresResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Búsqueda y Catálogo"
        ]
      }
    },
    "/api/v1/orders/preview": {
      "post": {
        "operationId": "AlojamientosController_previewOrder",
        "summary": "Previsualizar orden antes de confirmar",
        "description": "Congela el precio por 15 minutos.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderPreviewRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detalles de la orden previsualizada y precios finales",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPreviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida (ProblemDetails)"
          },
          "403": {
            "description": "Scope o rol insuficiente (ProblemDetails)"
          },
          "409": {
            "description": "Conflicto: ROOM_NO_LONGER_AVAILABLE, PRICE_CHANGED (+Retry-After)… (ProblemDetails)"
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "alojamientos:read"
            ]
          }
        ],
        "tags": [
          "Gestión de Órdenes (Reservas)"
        ]
      }
    },
    "/api/v1/orders/create": {
      "post": {
        "operationId": "AlojamientosController_createOrder",
        "summary": "Crear reserva de alojamiento",
        "description": "Pago simulado: payment_reference ^PAY-[A-Z0-9]{6,}$; PAY-DECLINED… se rechaza (402 PAYMENT_NOT_AUTHORIZED).",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID v4 para evitar cobros duplicados",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Orden creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida (ProblemDetails)"
          },
          "402": {
            "description": "PAYMENT_NOT_AUTHORIZED (ProblemDetails)"
          },
          "409": {
            "description": "Conflicto: ROOM_NO_LONGER_AVAILABLE, PRICE_CHANGED (+Retry-After)… (ProblemDetails)"
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "alojamientos:book"
            ]
          }
        ],
        "tags": [
          "Gestión de Órdenes (Reservas)"
        ]
      }
    },
    "/api/v1/orders/{orderId}": {
      "get": {
        "operationId": "AlojamientosController_getOrder",
        "summary": "Obtener detalles de la orden",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles completos de la orden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado (ProblemDetails)"
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "alojamientos:read"
            ]
          }
        ],
        "tags": [
          "Gestión de Órdenes (Reservas)"
        ]
      }
    },
    "/api/v1/orders/{orderId}/modify": {
      "post": {
        "operationId": "AlojamientosController_modifyOrder",
        "summary": "Modificar una orden existente",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID v4 para evitar modificaciones duplicadas",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderModifyRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Orden modificada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: ROOM_NO_LONGER_AVAILABLE, PRICE_CHANGED (+Retry-After)… (ProblemDetails)"
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "alojamientos:book"
            ]
          }
        ],
        "tags": [
          "Gestión de Órdenes (Reservas)"
        ]
      }
    },
    "/api/v1/orders/{orderId}/cancel": {
      "post": {
        "operationId": "AlojamientosController_cancelOrder",
        "summary": "Cancelar una orden",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID v4 para evitar cancelaciones duplicadas",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelación procesada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelacionResultado"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: ROOM_NO_LONGER_AVAILABLE, PRICE_CHANGED (+Retry-After)… (ProblemDetails)"
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "alojamientos:cancel"
            ]
          }
        ],
        "tags": [
          "Gestión de Órdenes (Reservas)"
        ]
      }
    },
    "/api/v1/webhooks": {
      "get": {
        "operationId": "AlojamientosController_listWebhooks",
        "summary": "Listar suscripciones",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Suscripciones activas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebhookSubscription"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "alojamientos:webhooks"
            ]
          }
        ],
        "tags": [
          "Webhooks"
        ]
      },
      "post": {
        "operationId": "AlojamientosController_createWebhook",
        "summary": "Registrar webhook",
        "description": "Solo https. El secreto (32 bytes) se muestra únicamente en esta respuesta.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSubscription"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "alojamientos:webhooks"
            ]
          }
        ],
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/api/v1/webhooks/{id}": {
      "delete": {
        "operationId": "AlojamientosController_deleteWebhook",
        "summary": "Eliminar suscripción",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado"
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "alojamientos:webhooks"
            ]
          }
        ],
        "tags": [
          "Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Posada EC · Alojamientos API",
    "description": "Implementación del contrato \"GDS Alojamientos Core API\" (rutas del YAML bajo /api/v1) y rutas propias (auth, cuenta, anfitrión, admin). El ownerId de cada reserva se infiere del `sub` del JWT. OAuth2 simplificado: el token se obtiene con POST /api/v1/auth/login y se envía como Bearer.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Este servidor"
    }
  ],
  "components": {
    "securitySchemes": {
      "OAuth2Security": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://auth.booking-hub.com/oauth2/authorize",
            "tokenUrl": "https://auth.booking-hub.com/oauth2/token",
            "scopes": {
              "alojamientos:read": "Leer información de alojamientos, disponibilidades y reservas",
              "alojamientos:book": "Crear y alterar reservas",
              "alojamientos:cancel": "Cancelar reservas",
              "alojamientos:webhooks": "Gestionar webhooks"
            }
          },
          "clientCredentials": {
            "tokenUrl": "https://auth.booking-hub.com/oauth2/token",
            "scopes": {
              "alojamientos:read": "Leer información de alojamientos, disponibilidades y reservas",
              "alojamientos:book": "Crear y alterar reservas",
              "alojamientos:cancel": "Cancelar reservas",
              "alojamientos:webhooks": "Gestionar webhooks"
            }
          }
        },
        "description": "OAuth2 del contrato. En este prototipo el token Bearer se emite en /api/v1/auth/login."
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SearchAccommodationResponse": {
        "type": "object",
        "properties": {
          "next_page": {
            "type": "string",
            "nullable": true
          },
          "request_id": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemBusqueda"
            }
          }
        },
        "required": [
          "next_page",
          "request_id",
          "data"
        ]
      },
      "AvailabilityRequestDto": {
        "type": "object",
        "properties": {
          "accommodation": {
            "type": "number",
            "description": "Código entero del alojamiento",
            "minimum": 1
          },
          "booker": {
            "$ref": "#/components/schemas/BookerDto"
          },
          "checkin": {
            "type": "string",
            "pattern": "FECHA"
          },
          "checkout": {
            "type": "string",
            "pattern": "FECHA"
          },
          "guests": {
            "$ref": "#/components/schemas/HuespedesDto"
          },
          "extras": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "extra_charges",
                "include_bundle_variants"
              ]
            }
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD"
            ],
            "pattern": "/^[A-Z]{3}$/"
          }
        },
        "required": [
          "accommodation",
          "booker",
          "checkin",
          "checkout",
          "guests"
        ]
      },
...
      "AvailabilityResponse": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/DatosDisponibilidad"
          }
        },
        "required": [
          "request_id",
          "data"
        ]
      },
...
      "BulkAvailabilityResponse": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisponibilidadAlojamiento"
            }
          }
        },
        "required": [
          "request_id",
          "data"
        ]
      },
...
    }
  }
}